requirements.txt
gremlinpython==3.4.0
tqdm
Owlready2
pyyaml
pandas
